cleanup and better background recog
